Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #11:

This patch reduces the compiler warning count on Win64 by an additional
80, as reported by VS2013. The patch also removes a number of unnecessary casts.

+++ sql/event_parse_data.cc
@@ -475,7 +475,7 @@ void
 Event_parse_data::report_bad_value(const char *item_name, Item *bad_item)
 {
   char buff[120];
-  String str(buff,(uint32) sizeof(buff), system_charset_info);
+  String str(buff, sizeof(buff), system_charset_info);
   String *str2= bad_item->fixed? bad_item->val_str(&str):NULL;
   my_error(ER_WRONG_VALUE, MYF(0), item_name, str2? str2->c_ptr_safe():"NULL");
 }
